upload file, relationship(mongoose)

+++ src/controllers/videoControllers.js
@@ -1,5 +1,5 @@
-import Video from "../models/Video"
-
+import Video from "../models/Video";
+import User from "../models/User";
 
 // const fakeUser = {
 //     username: "jyoh",
@@ -21,7 +21,7 @@ export const home = async(req, res) => {
 };
 export const watch = async(req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
     }
@@ -29,18 +29,26 @@ export const watch = async(req, res) => {
 }
 export const getEdit = async(req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle : `Edit: ${video.title}`, video});
 }
 export const postEdit = async(req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({_id:id});
     if(!video){
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title, 
@@ -53,22 +61,38 @@ export const getUpload = (req, res) => {
     return res.render("upload", {pageTitle : "upload video"});
 }
 export const postUpload = async (req, res) => {
+    const { user: { _id } } = req.session;
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
-            hashtags: Video.formatHashtags(hashtags)
+            hashtags: Video.formatHashtags(hashtags),
+            fileUrl,
+            owner: _id,
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         // const dbVideo = await video.save(); // promise이기 때문에 await 사용
         return res.redirect("/");
     }catch(error){
+        console.log(error);
         return res.status(400).render("upload", {pageTitle : "upload video", errorMessage: error._message});
     }
 }
 
 export const deletVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
